chore(Filterbar): move add/edit button in settings
