+++ superset-frontend/src/dashboard/components/nativeFilters/FilterBar/FilterBarSettings/index.tsx
@@ -39,6 +39,8 @@ import DropdownSelectableIcon, {
 import Checkbox from 'src/components/Checkbox';
 import { clearDataMaskState } from 'src/dataMask/actions';
 import { useCrossFiltersScopingModal } from '../CrossFilters/ScopingModal/useCrossFiltersScopingModal';
+import FilterConfigurationLink from '../FilterConfigurationLink';
+import { useFilters } from 'src/dashboard/components/nativeFilters/FilterBar/state';
 
 type SelectedKey = FilterBarOrientation | string | number;
 
@@ -65,11 +67,13 @@ const StyledCheckbox = styled(Checkbox)`
 
 const CROSS_FILTERS_MENU_KEY = 'cross-filters-menu-key';
 const CROSS_FILTERS_SCOPING_MENU_KEY = 'cross-filters-scoping-menu-key';
+const ADD_EDIT_FILTERS_MENU_KEY = 'add-edit-filters-menu-key';
 
 const isOrientation = (o: SelectedKey): o is FilterBarOrientation =>
   o === FilterBarOrientation.Vertical || o === FilterBarOrientation.Horizontal;
 
-const FilterBarSettings = () => {
+const FilterBarSettings = (
+) => {
   const dispatch = useDispatch();
   const theme = useTheme();
   const isCrossFiltersEnabled = useSelector<RootState, boolean>(
@@ -90,6 +94,11 @@ const FilterBarSettings = () => {
   );
   const canEdit = useSelector<RootState, boolean>(
     ({ dashboardInfo }) => dashboardInfo.dash_edit_perm,
+  );
+  const filters = useFilters();
+  const filterValues = useMemo(() => Object.values(filters), [filters]);
+  const dashboardId = useSelector<RootState, number>(
+    ({ dashboardInfo }) => dashboardInfo.id,
   );
   const canSetHorizontalFilterBar =
     canEdit && isFeatureEnabled(FeatureFlag.HorizontalFilterBar);
@@ -166,6 +175,16 @@ const FilterBarSettings = () => {
   const menuItems = useMemo(() => {
     const items: DropDownSelectableProps['menuItems'] = [];
 
+    if (canEdit){
+    items.push({
+      key: ADD_EDIT_FILTERS_MENU_KEY,
+      label: <FilterConfigurationLink
+      dashboardId= {dashboardId}
+      createNewOnOpen={filterValues.length === 0}
+    >{t('Add or edit filters')}
+    </FilterConfigurationLink>,
+      divider: canSetHorizontalFilterBar,
+    })}
     if (isCrossFiltersFeatureEnabled && canEdit) {
       items.push({
         key: CROSS_FILTERS_MENU_KEY,
@@ -177,7 +196,6 @@ const FilterBarSettings = () => {
         divider: canSetHorizontalFilterBar,
       });
     }
-
     if (canSetHorizontalFilterBar) {
       items.push({
         key: 'placement',
