+++ superset-frontend/src/dashboard/components/nativeFilters/FilterBar/FilterConfigurationLink/index.tsx
@@ -20,8 +20,6 @@ import { ReactNode, FC, useCallback, useState, memo } from 'react';
 
 import { useDispatch } from 'react-redux';
 import { setFilterConfiguration } from 'src/dashboard/actions/nativeFilters';
-import Button from 'src/components/Button';
-import { styled } from '@superset-ui/core';
 import FiltersConfigModal from 'src/dashboard/components/nativeFilters/FiltersConfigModal/FiltersConfigModal';
 import { getFilterBarTestId } from '../utils';
 import { SaveFilterChangesType } from '../../FiltersConfigModal/types';
@@ -33,10 +31,6 @@ export interface FCBProps {
   onClick?: () => void;
   children?: ReactNode;
 }
-
-const HeaderButton = styled(Button)`
-  padding: 0;
-`;
 
 export const FilterConfigurationLink: FC<FCBProps> = ({
   createNewOnOpen,
@@ -69,14 +63,12 @@ export const FilterConfigurationLink: FC<FCBProps> = ({
   return (
     <>
       {/* eslint-disable-next-line jsx-a11y/no-static-element-interactions */}
-      <HeaderButton
+      <span
         {...getFilterBarTestId('create-filter')}
-        buttonStyle="link"
-        buttonSize="xsmall"
         onClick={handleClick}
       >
         {children}
-      </HeaderButton>
+      </span>
       <FiltersConfigModal
         isOpen={isOpen}
         onSave={submit}
