+++ superset-frontend/src/dashboard/components/nativeFilters/FilterBar/Header/index.tsx
@@ -18,13 +18,9 @@
  */
 /* eslint-disable no-param-reassign */
 import { css, styled, t, useTheme } from '@superset-ui/core';
-import { memo, FC, useMemo } from 'react';
+import { memo, FC} from 'react';
 import Icons from 'src/components/Icons';
 import Button from 'src/components/Button';
-import { useSelector } from 'react-redux';
-import FilterConfigurationLink from 'src/dashboard/components/nativeFilters/FilterBar/FilterConfigurationLink';
-import { useFilters } from 'src/dashboard/components/nativeFilters/FilterBar/state';
-import { RootState } from 'src/dashboard/types';
 import { getFilterBarTestId } from '../utils';
 import FilterBarSettings from '../FilterBarSettings';
 
@@ -73,36 +69,9 @@ type HeaderProps = {
   toggleFiltersBar: (arg0: boolean) => void;
 };
 
-const AddFiltersButtonContainer = styled.div`
-  ${({ theme }) => css`
-    margin-top: ${theme.gridUnit * 2}px;
-
-    & button > [role='img']:first-of-type {
-      margin-right: ${theme.gridUnit}px;
-      line-height: 0;
-    }
-
-    span[role='img'] {
-      padding-bottom: 1px;
-    }
-
-    .ant-btn > .anticon + span {
-      margin-left: 0;
-    }
-  `}
-`;
-
 const Header: FC<HeaderProps> = ({ toggleFiltersBar }) => {
   const theme = useTheme();
-  const filters = useFilters();
-  const filterValues = useMemo(() => Object.values(filters), [filters]);
-  const canEdit = useSelector<RootState, boolean>(
-    ({ dashboardInfo }) => dashboardInfo.dash_edit_perm,
-  );
-  const dashboardId = useSelector<RootState, number>(
-    ({ dashboardInfo }) => dashboardInfo.id,
-  );
-
+  
   return (
     <Wrapper>
       <TitleArea>
@@ -117,16 +86,6 @@ const Header: FC<HeaderProps> = ({ toggleFiltersBar }) => {
           <Icons.Expand iconColor={theme.colors.grayscale.base} />
         </HeaderButton>
       </TitleArea>
-      {canEdit && (
-        <AddFiltersButtonContainer>
-          <FilterConfigurationLink
-            dashboardId={dashboardId}
-            createNewOnOpen={filterValues.length === 0}
-          >
-            <Icons.PlusSmall /> {t('Add/Edit Filters')}
-          </FilterConfigurationLink>
-        </AddFiltersButtonContainer>
-      )}
     </Wrapper>
   );
 };
